BAckend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
numpy==1.24.3
librosa==0.10.1
soundfile==0.12.1
torch==2.1.0
transformers==4.35.0
scipy==1.11.3
pandas==2.1.3
pyttsx3==2.90
gtts==2.4.0
pyaudio==0.2.13
wave==0.0.2
python-dotenv==1.0.0
cors==1.0.1
noisereduce==2.0.1

# Database Dependencies
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.0

# Additional Security
bcrypt==4.1.1
email-validator==2.1.0
